+++ file-sharing/vite.config.ts
@@ -2,7 +2,6 @@ import { fileURLToPath, URL } from "node:url";
 
 import { defineConfig, PluginOption } from "vite";
 import vue from "@vitejs/plugin-vue";
-import vueJsx from "@vitejs/plugin-vue-jsx";
 import VueDevTools from "vite-plugin-vue-devtools";
 
 import manifest from "../manifest.json";
@@ -13,7 +12,7 @@ const getAppVersionDefine = () => {
 
 // https://vitejs.dev/config/
 export default defineConfig({
-  plugins: [vue(), vueJsx(), VueDevTools()] as PluginOption[],
+  plugins: [vue(), VueDevTools()] as PluginOption[],
   base: "./",
   resolve: {
     alias: {